Refactoring
Every solution so far has had the tests pulled out in to a separate module, denoted by "test_". These test modules can be executed from command line as the source modules were previously. Pulled FileReader in to a utility.py module.
