+++ solutions/test_logs.py
@@ -1,0 +1,13 @@
+import unittest
+
+# class LogReaderTests:
+#
+#
+# class LogItemTests:
+#
+#
+# class LogSearchTests:
+
+
+if __name__ == '__main__':
+    unittest.main()
