+++ solutions/utility.py
@@ -1,0 +1,6 @@
+class FileReader:
+    def read_file(self, filename):
+        with open(filename, 'r') as f:
+            contents = f.read()
+
+        return contents
